requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
dnspython==2.7.0
frozenlist==1.4.1
idna==3.10
magic-filter==1.0.12
motor==3.6.0
MouseInfo==0.1.3
multidict==6.1.0
pillow==10.4.0
PyAutoGUI==0.9.54
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
pymongo==4.9.2
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
pytz==2024.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.3
yarl==1.13.1
